cli: share report UI test helpers via a testing module
make_account was copied verbatim into the app and event test modules,
and the fake-filesystem ledger construction plus the register-query
template were repeated across app, event, render, and the report
session-loop tests.

Add a #[cfg(test)] report::testing module holding the shared primitives
(template, fake_loader, process, make_account) and route the four test
modules through it, dropping the duplicates and their now-unused imports.
